🚚 : rename api service file for stacks
