+++ src/main/client/app/shared/api/stacks-api.js
